requirements.txt
crisp-api==1.0.12
boto3==1.35.99
openai==1.86.0
python-socketio[asyncio_client]==5.11.2
python-telegram-bot[all]==21.1.1
PyYAML==6.0.1
Requests==2.31.0

# 持久化相关依赖
# SQLite是Python内置的，无需额外安装
# 如果需要更高级的数据库功能，可以考虑添加：
# sqlalchemy>=2.0.0
# alembic>=1.8.0